wii-esc-ng:
Cosmetic changes

git-svn-id: https://wii-esc.googlecode.com/svn/branches/wii-esc-ng@211 5c42c5a8-900f-5b5a-2f8c-d720ba63385b

+++ src/hal/m8.h
@@ -30,8 +30,8 @@ void __delay_us(uint16_t __us) {
   while (__interval(i_start) < __us) {};
 };
 
-void __delay_ms(uint32_t __ms) __attribute__ ((noinline));
-void __delay_ms(uint32_t __ms) {
+void __delay_ms(uint16_t __ms) __attribute__ ((noinline));
+void __delay_ms(uint16_t __ms) {
   while (__ms--) __delay_us(1000);
 }
